First Commit
At this point there is an api to send a string of code to be executed on
a server running flask.
Currently 'run' is used as the delimiter between statements and
expressions.  That way the statements are exec'd and then the
expressions are eval'd.  stdout and stderr are both redirected and is
also returned with the output of each eval.